update task 2

+++ inno-de-lab_task_6/Task_2/main.py
@@ -1,8 +1,8 @@
 # input length of square
-length = float(input("Введите длину прямоугольника: "))
+length = int(input("Введите длину прямоугольника: "))
 
 # input width of square
-width = float(input("Введите ширину прямоугольника: "))
+width = int(input("Введите ширину прямоугольника: "))
 
 # output result
 print(f"Площадь прямоугольника: {length * width}")
